* Fixed a compiler warning when building with VS2010 64bit
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/branches/stable@181 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ core/platform/windows/pthread_create.h
@@ -550,7 +550,7 @@ static unsigned int WINAPI pthread_create_wrapper(void *args)
 static int pthread_create(pthread_t *th, pthread_attr_t *attr, void *(* func)(void *), void *arg)
 {
 	struct _pthread_v *tv = (struct _pthread_v *)malloc(sizeof(struct _pthread_v));
-	unsigned ssize = 0;
+	size_t ssize = 0;
 	
 	if (!tv) return 1;
 	
@@ -575,7 +575,7 @@ static int pthread_create(pthread_t *th, pthread_attr_t *attr, void *(* func)(vo
 	/* Make sure tv->h has value of -1 */
 	_ReadWriteBarrier();
 
-	tv->h = (HANDLE) _beginthreadex(NULL, ssize, pthread_create_wrapper, tv, 0, NULL);
+	tv->h = (HANDLE) _beginthreadex(NULL, (unsigned int)ssize, pthread_create_wrapper, tv, 0, NULL);
 	
 	/* Failed */
 	if (!tv->h) return 1;
